Remove elements that occur before the first bin when splitting by bins
This change is in-line with the current treatment of the elements that
occur after the last bin.

Further, this fixes a minor "bug" where bin indecees are used to
determine bin labels by indexing into a list of labels. Elements before
the first bin received bin index 0, however, R is 1-indexed. Note: This
"bug" did not cause problems because the bin with the incorrect label
will be ignored later anyways.

Works towards #267.

Signed-off-by: Maximilian Löffler <s8maloef@stud.uni-saarland.de>